added seeder in database seeder
