+++ app/Http/Controllers/Formation/FormationController.php
@@ -1,18 +1,17 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Formation;
 
-use App\Models\JobOffer;
-use App\Models\Skill;
+use App\Http\Controllers\Controller;
+use App\Models\Formation;
 use Illuminate\Http\Request;
 
-class JobOfferController extends Controller
+class FormationController extends Controller
 {
-
     public function index()
     {
-        $joboffers = JobOffer::all();
-        return view("", compact("joboffers"));
+        $formations = Formation::all();
+        return view("", compact("formations"));
     }
 
     /**
@@ -29,7 +28,7 @@ class JobOfferController extends Controller
     public function store(Request $request)
     {
         //
-        JobOffer::create($request->all());
+        Formation::create($request->all());
         return redirect()->route("");
     }
 
@@ -45,19 +44,19 @@ class JobOfferController extends Controller
     /**
      * Update the specified resource in storage.
      */
-    public function update(Request $request , JobOffer $jobOffer)
+    public function update(Request $request , Formation $formation)
     {
         //
-        $jobOffer->update($request->all());
+        $formation->update($request->all());
         return redirect()->route("");
     }
 
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(JobOffer $jobOffer)
+    public function destroy(Formation $formation)
     {
-        $jobOffer->delete();
+        $formation->delete();
         return redirect()->route("");
     }
