+++ app/Http/Controllers/Skill/SkillController.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Skill;
 
-use App\Models\Formation;
+use App\Http\Controllers\Controller;
 use App\Models\Skill;
 use App\Models\User;
 use Illuminate\Http\Request;
