+++ resources/views/candidat/candidat_profile.blade.php
@@ -7,15 +7,15 @@
 </head>
 <body>
 
+
+@include("components.Navbar")
+
 <div class="container mt-5">
     <section style="background-color: #eee;">
         <div class="container py-5" style="width: 100%">
             <div class="row">
                 <div class="col-lg-4">
                     <div class="card mb-4">
-
-
-
                         <div class="card-body text-center">
                                 <img src="{{$profile->getFirstMediaUrl('profile_images')}}" alt="avatar" class="rounded-circle img-fluid">
                             <h5 class="my-3">{{ $profile->name }}</h5>
@@ -38,19 +38,19 @@
                             <ul class="list-group list-group-flush rounded-3">
                                 <li class="list-group-item d-flex justify-content-between align-items-center p-3">
                                     <i class="fas fa-globe fa-lg text-warning"></i>
-                                    <p class="mb-0">https://mdbootstrap.com</p>
+                                    <p class="mb-0">{{$profile->company->name}}</p>
                                 </li>
                                 <li class="list-group-item d-flex justify-content-between align-items-center p-3">
                                     <i class="fab fa-github fa-lg" style="color: #333333;"></i>
-                                    <p class="mb-0">mdbootstrap</p>
+                                    <p class="mb-0">{{$profile->company->description}}</p>
                                 </li>
                                 <li class="list-group-item d-flex justify-content-between align-items-center p-3">
                                     <i class="fab fa-twitter fa-lg" style="color: #55acee;"></i>
-                                    <p class="mb-0">@mdbootstrap</p>
+                                    <p class="mb-0">{{$profile->company->name}}</p>
                                 </li>
                                 <li class="list-group-item d-flex justify-content-between align-items-center p-3">
                                     <i class="fab fa-instagram fa-lg" style="color: #ac2bac;"></i>
-                                    <p class="mb-0">mdbootstrap</p>
+                                    <p class="mb-0">{{$profile->company->city->name}}</p>
                                 </li>
                                 <li class="list-group-item d-flex justify-content-between align-items-center p-3">
                                     <i class="fab fa-facebook-f fa-lg" style="color: #3b5998;"></i>
